More

+++ src/bioregistry/external/miriam/__init__.py
@@ -6,10 +6,12 @@ from operator import itemgetter
 from pathlib import Path
 from typing import Any, ClassVar
 
+from pydantic import BaseModel, Field
 from pystow.utils import download
 
+from bioregistry.alignment_model import Record, Status, dump_records, load_processed
 from bioregistry.constants import RAW_DIRECTORY, URI_FORMAT_KEY
-from bioregistry.external.alignment_utils import Aligner, load_processed
+from bioregistry.external.alignment_utils import Aligner
 
 __all__ = [
     "MiriamAligner",
@@ -35,21 +37,34 @@ SKIP_URI_FORMATS = {
 }
 
 
+class MiriamExtra(BaseModel):
+    """Extras in MIRIAM."""
+
+    namespace_embedded_in_lui: bool = Field(alias="namespaceEmbeddedInLui")
+    miriam_id: str
+
+
 def get_miriam(
     force_download: bool = False, force_process: bool = False
-) -> dict[str, dict[str, Any]]:
+) -> dict[str, Record[MiriamExtra]]:
     """Get the MIRIAM registry."""
     if PROCESSED_PATH.exists() and not force_download and not force_process:
         return load_processed(PROCESSED_PATH)
 
     download(url=MIRIAM_URL, path=RAW_PATH, force=force_download)
-    with open(RAW_PATH) as file:
-        data = json.load(file)
 
-    data["payload"]["namespaces"] = sorted(data["payload"]["namespaces"], key=itemgetter("prefix"))
-    if force_download:
-        with open(RAW_PATH, "w") as file:
-            json.dump(data, file, indent=2, sort_keys=True, ensure_ascii=False)
+    if RAW_PATH.is_file() and not force_process:
+        data = json.loads(RAW_PATH.read_text())
+    else:
+        with open(RAW_PATH) as file:
+            data = json.load(file)
+
+        data["payload"]["namespaces"] = sorted(
+            data["payload"]["namespaces"], key=itemgetter("prefix")
+        )
+        if force_download:
+            with open(RAW_PATH, "w") as file:
+                json.dump(data, file, indent=2, sort_keys=True, ensure_ascii=False)
 
     rv = {
         record["prefix"]: _process(record)
@@ -57,8 +72,7 @@ def get_miriam(
         # records whose prefixes start with `dg.` appear to be unreleased
         if not record["prefix"].startswith("dg.") and record["prefix"] not in SKIP
     }
-    with PROCESSED_PATH.open("w") as file:
-        json.dump(rv, file, indent=2, sort_keys=True)
+    dump_records(rv, PROCESSED_PATH)
     return rv
 
 
@@ -72,15 +86,17 @@ PROVIDER_BLACKLIST = {
 }
 
 
-def _process(record: dict[str, Any]) -> dict[str, Any]:
+def _process(record: dict[str, Any]) -> Record[MiriamExtra]:
     prefix = record["prefix"]
     rv = {
         "prefix": prefix,
-        "id": record["mirId"][len("MIR:") :],
         "name": record["name"],
-        "deprecated": record["deprecated"],
-        "namespaceEmbeddedInLui": record["namespaceEmbeddedInLui"],
-        "sampleId": record["sampleId"],
+        "status": Status.deprecated if record["deprecated"] else Status.active,
+        "extras": {
+            "namespaceEmbeddedInLui": record["namespaceEmbeddedInLui"],
+            "miriam_id": record["mirId"][len("MIR:") :],
+        },
+        "example": [record["sampleId"]],
         "description": record["description"],
         "pattern": record["pattern"],
     }
@@ -90,7 +106,7 @@ def _process(record: dict[str, Any]) -> dict[str, Any]:
         if not resource.get("deprecated")
     ]
     if not resources:
-        return rv
+        return Record[MiriamExtra].model_validate(rv)
 
     has_official = any(resource["official"] for resource in resources)
     if has_official:
@@ -116,7 +132,7 @@ def _process(record: dict[str, Any]) -> dict[str, Any]:
         extras.append(provider)
     if extras:
         rv["providers"] = extras
-    return rv
+    return Record[MiriamExtra].model_validate(rv)
 
 
 SKIP_PROVIDERS = {
